int-as-pointer parameters: 29 more slots, and the trigger flags read through their field
Same class C98 opened and C99 continued: a parameter declared int/u32 whose body
immediately casts it to a pointer.  ABI-identical, so no gate in this tree can
see it -- it is pure semantic debt.  Each component lands whole: definition,
every prototype, every call site, and the handler-table field where a DLL export
carries the slot.

  ObjTrigger (3 slots, 39 call sites)  ObjTrigger_IsSet, ObjTrigger_IsSetById
      and ObjTrigger_UpdateIdBlockFlag all take GameObject*.  IsSetById and
      UpdateIdBlockFlag were reaching the trigger byte as
      *(u8*)(obj + OBJTRIGGER_FLAGS_OFFSET); OBJTRIGGER_FLAGS_OFFSET is 0xAF and
      offsetof(ObjAnimComponent, resetHitboxFlags) == 0xAF is STATIC_ASSERTed in
      objanim_internal.h, so those five raw byte accesses are now
      obj->anim.resetHitboxFlags -- which is how the sibling ObjTrigger_IsSet
      already read the same byte.  Thirty-nine (int)/(u32) casts drop at the call
      sites; five sites keep a cast because the CALLER's own local is still an
      int (626's `register int self`, SHthorntail's four), and that local's debt
      is a different slot.
  DIM boss (3)  returnToIdleWhenDone, liftImpact and updateBossHitReaction, plus
      the three DIMbossAnimHandlerTable / DIMbossHitDetectAnimHandlerTable fields
      that carry them -- the other nine fields already said GameObject*/
      BaddieState*.  liftImpact's (ObjAnimComponent*)obj is &obj->anim, GameObject
      declaring anim first.
  Twenty-three single-slot DLL entry points whose descriptor-table registration
  already casts: 552 x2, 376, 223, 244, 640, 689, 614, 610, 577, 701 x2, 344,
  407, 650, 556, 551, 546, 432, 483, 482, 11.

Two are typed void*, not GameObject*, because that is what the evidence says and
nothing more: sky2ApplyTextColor's argument goes only to setTextColor(void*
context, ...) -- whose other callers pass &gdl, a display list, not an object --
so it takes the callee's own parameter name; dll_0B_beginSequence's goes only
into ExpgfxSpawnConfig.attachedSource, which is void*.

Measured and NOT landed, each reverted whole:
  WallAnimator_init  .text differs.  The parameter is used both as a pointer and
      as an int (objAddObjectType), and once it IS a pointer MWCC coalesces the
      web and re-colours -- C98's screen, in the form where there is no alias to
      delete.
  dll_200_update  .text grows 4 bytes for the same reason.
  earthwalker_update  .text differs; its sibling earthwalker_SeqFn is byte-equal
      and is landed.
  worldobj_render's renderArg2 and waterfx_render's obj  DECLINED, not reverted
      for byte reasons: each has two callees that disagree about the type
      (objRenderModelAndHitVolumes takes int, Camera_ApplyCurrentViewport takes
      void*; Camera_LoadModelViewMatrix takes int, setTextColor takes void*).
      Two roles in one slot, so no re-typing is justified.

Not touched: player.c, 597.c, objects/202, objprint.c, engine/21 (other lanes),
and the shared int-taking APIs those lanes' files call -- objAddObjectType/
objFreeObjectType (280 sites), ObjAnim_SetCurrentMove (492),
Sfx_KeepAliveLoopedObjectSound (60) -- which cannot close without them.

GATE, twice, build-idle, against a pristine control built at the same parent:
1005 objects compared / 0 differ / 0 missing / 0 new; 6 carry only @NNN
literal-pool renumbering at preserved addresses, which never reaches the link.
Tree unchanged at fuzzy 99.81484 / matched_code 2513248 / matched_data 1195969 /
complete_units 914 of 1043.  banned_shapes 103/84/21 identical to the control.
ninja all_source EXIT=0; forced link main.dol: OK, sha1
e750e8e894707a52446118a4b84f1b58b677b269 md5 7b955850ea4bd7ceda0109493203ff5b.

+++ src/dlls/objects/237/237.c
@@ -451,7 +451,7 @@ void collectible_checkProximityPickup(GameObject* obj, CollectibleState* state) 
             state->pickupLatch |= COLLECTIBLE_PICKUP_LATCHED;
             break;
         default:
-            if (ObjTrigger_IsSet((int)obj) != 0) {
+            if (ObjTrigger_IsSet(obj) != 0) {
                 mainSetBits(GAMEBIT_EnableCMenu, 1);
                 state->pickupMsgValue = placement->collectGameBit;
                 ObjMsg_SendToObject(player, COLLECTIBLE_MSG_IN_RANGE, (void*)obj,
